etl-solution/extractor/requirements.txt
h11==0.8.1
uvicorn==0.11.3
Werkzeug==1.0.1
sanic-restful-api==0.0.8
sanic-envconfig==1.3.3
confluent-kafka==1.4.0
jsonschema==3.2.0
json-logging==1.1.0
py-money==0.5.0
pytest==5.4.1